Update: Start init update field and rename hook
