+++ src/components/FieldModal/FieldModal.module.css
@@ -12,6 +12,21 @@
   background: rgba(0, 0, 0, 0.6);
 }
 
+.modalUpdateContainer {
+  position: absolute;
+  top: 0;
+  left: 0;
+  margin: 0;
+  padding: 0;
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: rgba(0, 0, 0, 0.6);
+  z-index: 10;
+}
+
 .containerForm {
   display: flex;
   flex-direction: column;
@@ -38,6 +53,7 @@
   height: 48px;
   top: 25%;
   left: 71%;
+  cursor: pointer;
 }
 
 .infosContainer, .optionsContainer {
